LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-center-fallback-transition-behavior-expected.txt


FAIL anchor-center aligned bottom anchored element overflowing IMCB assert_equals: Positioned using --right fallback expected 100 but got 0
